# Improve the speed of Polynomial::EvaluatePartial