Resolves #17700

Also add benchmark Polynomial::EvaluatePartial.

Before this change
```
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
BenchmarkPolynomialEvaluatePartial  731970787 ns    731922688 ns            1
-----------------------------------------------------------------------------
```
After this change
```
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
BenchmarkPolynomialEvaluatePartial    2896414 ns      2896364 ns          246
```

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/17701)
<!-- Reviewable:end -->