@jwnimmer-tri 2022-08-06
*[`common/symbolic/benchmarking/benchmark_polynomial.cc` line 47 at r1](https://reviewable.io/reviews/robotlocomotion/drake/17701#-N8nzYQ6A948bKZZVP1D:-N8nzYQ70ufbw7hMLlYr:b-eoqpn5) ([raw file](https://github.com/robotlocomotion/drake/blob/97d504f09248f5fa3fe893113926f7d0aacf2a76/common/symbolic/benchmarking/benchmark_polynomial.cc#L47)):*
> ```cpp
> }
> 
> BENCHMARK(BenchmarkPolynomialEvaluatePartial);
> ```

nit For the slower benchmarks, IMO reporting nanoseconds in the output is difficult to understand.  Consider adding `->Unit(benchmark::kMicrosecond)` here for readability.


<!-- Sent from Reviewable.io -->