@jwnimmer-tri 2022-08-10
Could you split a PR just to add the benchmark first, with no fixes?

I would like to dig into why the Substitute is faster, but I won't have time this week.

My intuition is that we land the benchmark first, then my patch next.  And then try to diagnose why Substitute is faster, and fix EvaluatePartial to be equally fast.