@jwnimmer-tri 2022-08-15
Merging this branch up to `upstream/master` and running locally as  `bazel run //common/benchmarking:polynomial_experiment -- --output_dir=... --sleep 1 --cputask=4 -- --benchmark_repetitions=30`, with this change I see 4778±68 us measured CPU time and with #17727 I see 4049±64 us measured CPU time.  So on my computer, there is no 2x slower for #17727 vs this approach.

Could you share how you took your measurements?  We you using `polynomial_experiment`?  If you were running `benchmark_polynomial` directly, you'd be subject to CPU turbo scaling, which makes measurements inaccurate below ~5x relative scale.